% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmsc.R
\name{lmsc}
\alias{lmsc}
\title{Localized Multiplicative Scatter Correction (LMSC)}
\usage{
lmsc(X, k, r)
}
\arguments{
\item{X}{row vector or matrix of spectra to correct}

\item{k}{number of segments}

\item{r}{reference spectrum}
}
\value{
Matrix of LMSC-corrected spectra
}
\description{
Perform localized multiplicative scatter correction (LMSC), i.e., MSC performed separately on k disjoint but continuous spectral segments.
}
\examples{
data("marzipan")
X <- marzipan$X
ref <- colMeans(X)
X.lmsc <- lmsc(X,5,ref) #LMSC with 5 segments
matplot(as.vector(marzipan$ax),t(X.lmsc),type="l",lty=1)
}
\references{
C. Giglio, J.-M. Roger, E. Andries. Piecewise Standard Normal Variate (PSNV) for scatter correction. J Chemometrics (2025) (Submitted)

C.A. Andersson, Chemometrics and Intelligent Laboratory Systems 1999, 47 (1), 51–63.
}
