% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnv.R
\name{lsnv}
\alias{lsnv}
\title{Localized Standard Normal Variate (LSNV)}
\usage{
lsnv(X, k)
}
\arguments{
\item{X}{row vector or matrix of spectra to correct}

\item{k}{number of segments}
}
\value{
Matrix of LSNV-corrected spectra
}
\description{
Perform localized standard normal variate (LSNV), i.e., SNV performed separately on k disjoint but continuous spectral segments.
}
\examples{
#data("marzipan")
#X.lsnv <- lsnv(marzipan$X,5) #LSNV with 5 segments
#matplot(as.vector(marzipan$ax),t(X.lsnv),type="l",lty=1)
}
\references{
C. Giglio, J.-M. Roger, E. Andries. Piecewise Standard Normal Variate (PSNV) for scatter correction. J Chemometrics (2025) (Submitted)

Y. Bi, K. Yuan, W. Xiao, J. Wu, C. Shi, J. Xia, G. Chu, G. Zhang, G. Zhou, Analytica chimica acta 2016, 909, 30–40.
}
