% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movsum.R
\name{movsum}
\alias{movsum}
\title{Moving Sum}
\usage{
movsum(X, w)
}
\arguments{
\item{X}{matrix of data, dimension m x n. If X is a vector, convert to a
row matrix using matrix(X,nrow=1)}

\item{w}{window size, an odd integer. If we define the half-window size as v = (w-1)/2, and ncol(X)=n, then the moving window at channel j will span from
max(1,j-v) to min(n,j+v). In other words, the window is symmetric for the middle columns, and is asymmetric at the left and right edges.}
}
\value{
Matrix of data with moving standard deviation corrected data
}
\description{
Wrapper function for C++ code to evaluate a moving sum over each row for a matrix X. Equivalent to MATLAB's 'movsum' function.
}
\examples{
set.seed(1234)
#X <- matrix(sample(-10:10,3*10,replace=T),nrow=3,ncol=10) #generate synthetic data; equivalent to the line below:
X <- rbind(c(5,4,-5,-9,4,3,10,-8,-9,9),c(-6,-2,5,-4,3,-7,-3,-7,4,5),c(1,-6,-7,-5,9,-7,9,-6,-3,1))
#evaluate a 5-point moving sum on each row of X:
X_msd <- movsum(X,3)

If we load X into MATLAB, we can verify that it is equivalent (within machine tolerance) to:
x1 = [5,4,-5,-9,4,3,10,-8,-9,9];
x2 = [-6,-2,5,-4,3,-7,-3,-7,4,5];
x3 = [1,-6,-7,-5,9,-7,9,-6,-3,1];
X = [x1;x2;x3];
X_msd = movsum(X,5,2);

}
