% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSC.R
\name{MSC}
\alias{MSC}
\title{Multiplicative Scatter Correction (MSC)}
\usage{
MSC(X, r = colMeans(X))
}
\arguments{
\item{X}{matrix of spectra to correct (if X is a vector, use matrix(X,nrow=1) to convert to a row matrix)}

\item{r}{reference spectrum (default: average of each column of X)}
}
\value{
Matrix of MSC-corrected spectra
}
\description{
Performs multiplicative scatter correction (MSC). The code is a copy of the function 'msc' from the package 'pls', but has been slightly optimized for PMSC and LMSC
For standalone use, the code has no input sanitization or checks, so please ensure that class(X)="matrix" and that length(ref)=ncol(X) before use
}
\examples{
data("marzipan")
ref <- colMeans(marzipan$X)
X.msc <- MSC(marzipan$X,ref) #PMSC with window size of 101 channels
matplot(marzipan$ax,t(X.msc),type="l",lty=1)
}
\references{
C. Giglio, J.-M. Roger, E. Andries. Piecewise Standard Normal Variate (PSNV) for scatter correction. J Chemometrics (2025) (Submitted)
P Geladi, D MacDougall, H Martens, Applied Spectroscopy 1985, 39 (3), 491–500.
}
