% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmsc.R
\name{pmsc}
\alias{pmsc}
\title{Piecewise Multiplicative Scatter Correction (PMSC)}
\usage{
pmsc(X, w, r, method = c("cpp", "simple", "mpmsc"))
}
\arguments{
\item{X}{row vector or matrix of spectra to correct}

\item{w}{moving window size (must be an odd integer)}

\item{r}{reference spectrum}

\item{method}{Algorithm to calculate PSMC correction (see below for details)}
}
\value{
Matrix with PMSC-corrected spectra
}
\description{
Perform piecewise multiplicative scatter correction (PMSC), i.e., MSC performed separately on a sliding band centered at the jth column of X
}
\details{
3 different algorithms for PMSC are supported using the 'methods' argument: "cpp" (default), "simple", and "mpmsc":

"simple" will calculate the full MSC correction at a given moving window, and will extract the correction corresponding to the j-th wavelength.

"cpp" calculates the MSC coefficients corresponding to the moving window surrounding the j-th column of X, and performs the correction using this manner.
This is all calculated inside a C++ wrapper function for speedup. This version is the fastest to calculate.

"mpmsc" uses moving sums and moving means to calculate the correction; This version is optimal in MATLAB but is slightly slower in R
}
\examples{
data("marzipan")
ref <- colMeans(marzipan$X)
X.pmsc <- mpmsc(marzipan$X,201,ref) #PMSC with window size of 201 channels
matplot(as.vector(marzipan$ax),t(X.pmsc),type="l",lty=1)
}
\references{
C. Giglio, J.-M. Roger, E. Andries. Piecewise Standard Normal Variate (PSNV) for scatter correction. J Chemometrics (2025) (Submitted)

T. Isaksson, B. Kowalski, Applied Spectroscopy 1993, 47 (6), 702–709
}
