% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psnv.R
\name{psnv}
\alias{psnv}
\title{Piecewise Standard Normal Variate (PSNV)}
\usage{
psnv(X, w, method = c("fast", "simple"))
}
\arguments{
\item{X}{vector or matrix of spectra to correct}

\item{w}{moving window size (must be an odd integer)}

\item{method}{Algorithm to calculate PSNV correction (see below for details)}
}
\value{
Matrix with PSNV-corrected spectra
}
\description{
Calculate piecewise standard normal variate correction (PSNV), i.e. SNV evaluated using a rolling average and rolling standard deviation.
For a window size w (an odd integer between 3 and 2:ncol(X)), the half-window size v is (w-1)/2. The moving window for channel j is equal to
max(1,j-v) to min(j+v,ncol(X)). As a result, at the edges of the window are asymmetric at the edges of the spectrum
}
\details{
Two different implementations are included for PSNV: "fast" and "simple"
The fast algorithm uses C++ functions to evaluate the rolling average and rolling standard deviation, and is significantly more efficient
The simple algorithm calculates the mean and standard deviation of the local window corresponding to each column, and is far slower to calculate
}
\examples{
data("marzipan")
X.psnv <- psnv(marzipan$X,101) #PSNV with window size of 101 channels
matplot(as.vector(marzipan$ax),t(X.psnv),type="l",lty=1)


}
\references{
C. Giglio, J.-M. Roger, E. Andries. Piecewise Standard Normal Variate (PSNV) for scatter correction. J Chemometrics (2025) (Submitted)

A. Rinnan, F. van den Berg, SB Engelsen, TrAC Trends in Analytical Chemistry, 2009, 28(10), 1201-1222. (Earliest paper to mention this method)
}
\author{
Cannon Giglio
}
